first draft of features
