+++ Cargo.toml
@@ -20,18 +20,19 @@ version.workspace = true
 workspace = true
 
 [features]
-default = ["client"]
-# default = ["client", "server"]
+# default = ["client"]
+default = ["client", "server"]
 # Includes DWN client features such as request builders.
 client = []
 # Includes DWN server features such handlers.
-server = []
+server = ["dep:async-recursion", "dep:tokio"]
 
 [dependencies]
 aes-gcm = "0.10.3"
 anyhow.workspace = true
-async-recursion = "1.1.1"
+async-recursion = { version = "1.1.1", optional = true }
 base64ct.workspace = true
+cfg-if = "1.0.0"
 chrono = { version = "0.4.39", features = ["serde"] }
 ciborium = "0.2.2"
 cid.workspace = true
@@ -51,7 +52,7 @@ serde_ipld_dagcbor = "0.6.1"
 serde_json.workspace = true
 sha2.workspace = true
 thiserror = "2.0.11"
-tokio.workspace = true
+tokio = { workspace = true, optional = true }
 ulid = "1.2.0"
 uuid = { version = "1.13.1", features = ["fast-rng", "v4"] }
 x25519-dalek.workspace = true
@@ -59,7 +60,6 @@ zeroize = "1.8.1"
 
 [dev-dependencies]
 blockstore = "0.7.1"
-insta.workspace = true
 test-node.workspace = true
 
 [workspace]
@@ -100,7 +100,6 @@ dwn-node = { path = "." }
 ed25519-dalek = { version = "2.1.1", features = ["rand_core"] }
 futures = "0.3.31"
 http = "1.2.0"
-insta = { version = "1.42.1", features = ["redactions", "yaml"] }
 multibase = "0.9"
 rand = "0.8.5"
 serde = { version = "1.0.217", features = ["derive"] }
@@ -114,7 +113,7 @@ x25519-dalek = { version = "2.0.1", features = ["static_secrets"] }
 # credibil-infosec = { path = "../infosec" }
 
 [profile.release]
-# Optimize for small code size
+# optimize for small code size
 lto = true
 opt-level = "s"
 strip = true
