+++ crates/test-node/Cargo.toml
@@ -13,13 +13,12 @@ async-nats = "0.38.0"
 base64ct.workspace = true
 blockstore = "0.7.1"
 cid.workspace = true
-dwn-node.workspace = true
+credibil-infosec.workspace = true
+dwn-node = { workspace = true, features = ["server"] }
 ed25519-dalek.workspace = true
 futures.workspace = true
-insta.workspace = true
 multibase.workspace = true
 rand.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
-credibil-infosec.workspace = true
 x25519-dalek.workspace = true
